+++ contracts/cosmwasm-vm/cw-ibc-core/src/state.rs
@@ -13,7 +13,7 @@ pub struct CwIbcStore<'a> {
     client_connections: Map<'a, ClientId, ConnectionId>,
     connections: Map<'a, ConnectionId, Vec<u8>>,
     channels: Map<'a, (PortId, ChannelId), ChannelEnd>,
-    port_to_moudle: Map<'a, PortId, ModuleId>,
+    port_to_moudle: Map<'a, PortId, IbcModuleId>,
     /// Stores address based on the capabilty names
     capabilities: Map<'a, Vec<u8>, Vec<String>>,
     /// store commitments based on keys (PacketCommitment,AckCommitment,Connection,Channel,Client)
@@ -90,7 +90,7 @@ impl<'a> CwIbcStore<'a> {
     pub fn channels(&self) -> &Map<'a, (PortId, ChannelId), ChannelEnd> {
         &self.channels
     }
-    pub fn port_to_moulde(&self) -> &Map<'a, PortId, ModuleId> {
+    pub fn port_to_moulde(&self) -> &Map<'a, PortId, IbcModuleId> {
         &self.port_to_moudle
     }
     pub fn capabilities(&self) -> &Map<'a, Vec<u8>, Vec<String>> {
